+++ testsuite/docker/kafka/scripts/start_no_wait.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 set -e
 
-[ "$KAFKA_ZOOKEEPER_CONNECT" == "" ] && KAFKA_ZOOKEEPER_CONNECT=localhost:2181
-[ "$KAFKA_ZOOKEEPER_CONNECTION_TIMEOUT_MS" == "" ] && KAFKA_ZOOKEEPER_CONNECTION_TIMEOUT_MS=6000
-
 ./simple_kafka_config.sh | tee /tmp/strimzi.properties
+echo "Config created"
+
+KAFKA_CLUSTER_ID="$(/opt/kafka/bin/kafka-storage.sh random-uuid)"
+/opt/kafka/bin/kafka-storage.sh format -t $KAFKA_CLUSTER_ID -c /tmp/strimzi.properties
+echo "Initialised kafka storage for KRaft"
 
 # set log dir to writable directory
 if [ "$LOG_DIR" == "" ]; then