+++ testsuite/keycloak-errors-tests/src/test/java/io/strimzi/testsuite/oauth/auth/ErrorReportingTests.java
@@ -23,6 +23,7 @@ import java.util.concurrent.ExecutionException;
 import static io.strimzi.kafka.oauth.common.OAuthAuthenticator.loginWithClientSecret;
 import static io.strimzi.testsuite.oauth.auth.Common.buildProducerConfigOAuthBearer;
 import static io.strimzi.testsuite.oauth.auth.Common.buildProducerConfigPlain;
+import static io.strimzi.testsuite.oauth.common.TestUtil.assertTrueExtra;
 import static io.strimzi.testsuite.oauth.common.TestUtil.getContainerLogsForString;
 import static io.strimzi.testsuite.oauth.common.TestUtil.getRootCause;
 
@@ -463,9 +464,10 @@ public class ErrorReportingTests {
             Assert.fail("Should fail with KafkaException");
         } catch (Exception e) {
             long diff = System.currentTimeMillis() - start;
-            Assert.assertTrue("is instanceof KafkaException", e instanceof KafkaException);
-            Assert.assertTrue("Failed due to LoginException", getRootCause(e).toString().contains("LoginException"));
-            Assert.assertTrue("Unexpected diff: " + diff, diff > timeout * 1000 && diff < timeout * 1000 + 1000);
+
+            assertTrueExtra("is instanceof KafkaException", e instanceof KafkaException, e);
+            assertTrueExtra("Failed due to LoginException", getRootCause(e).toString().contains("LoginException"), e);
+            assertTrueExtra("Unexpected diff: " + diff, diff > timeout * 1000 && diff < timeout * 1000 + 1000, e);
         }
     }